feat: Add query timing data to result (#21)
* feat: Add query timing data to result

A simple decorator that provides a timer object to the decorated
function. The timer can be added to the result and will be serialized as
a a json object with timing data. Additionally timer.mark() can be used
at various points throughout the function to provide more fine grained
timing data for various steps.