fix: Expand the issues JOIN if issues are used anywhere.
Even if a query passes an empty `issues` mapping, we still need
to expand this JOIN if `issue` is referenced in the groupby or
conditions of the query to avoid a syntax error when 'issue' is not
defined.

In cases where `issues` is empty and `issue` is not referenced, we
can continue to just leave it out.