fix: Workaround for CH exception on `issue IN (x)`
When there are no issues defined, we still need to expand our join
expression so that conditions on `issue` dont become syntax errors.
We did this by joining to a single row of
`(issue, hash, tombstone) = (Null, Null, Null)`

However, for `IN` comparisons, the type of the `issue` column still
needs to be an int to avoid `Exception: Type mismatch in IN or VALUES
section. Expected: Nothing. Got: UInt64.` .

So now the null row in the join table is `(0, Null, Null)`