fix: Invert alias cache
Instead of re-using an alias when the generated _expression_ is the same
as a previously generated one, simply re-use the alias if we have ever
aliased _any_ expression to that same alias before.

This makes alias reuse simpler, but has the slight pitfall that if you
mess up and try and alias 2 expressions to the same thing, you will
silently get 2 copies of the first thing, rather than a Clickhouse
exception that you are trying to alias different expressions to the
same thing.