test: Test to see if only hitting the tags column helps
Before actually doing a full backfill with all the tags in the nested
column, check that this is actually a potential perf win by running
existing queries against just the tags field.

This will be very bad for accuracy results, as the top tags method
won't return any of the promoted tags. But I won't leave it running for
long