feat: Leave promoted tags in the nested tags list (#92)
Based on the observation that the tags_key expression to grab the union
of promoted and nested tags creates exceptionally slow queries, and
requires a bunch of other translation hacks etc. Instead we can just
leave the promoted tags in the nested tags column, and use only that
column when doing queries for all tags.

Note that queries for specific named tags eg. `tags[environment]` will
still use the promoted column to maintain optimal performance in the
case where we know which tag we are looking for.