Respect and prefer new canonical forms of interface keys.
Related to https://github.com/getsentry/sentry/pull/8789

This is the bare minimum we need so Snuba is forward-compatible with
events it may see going forward. There may be other coordination work or
schemas we share in the future, but this is here so we don't block
process on canonicalization.