fix: Parse Nullable(DateTime) columns
This was causing a json.dumps error when serializing datetimes in the result.

An alternative is just to use a JSON serializer that serializes datetimes,
but for now I think we want to retain control over which result columns
we are doing this for, so we don't just blindly jsonify weird objects we're
not expecting to the clients.