feat: Special endpoint for SDK stats (#173)
Add a custom endpoint for internal use that constructs a snuba query to
grab sdk stats across all projects, thus avoiding the standard
resttictions (defined set of project conditions) that apply to the
regular query endpoint. I'm not 100% sure this is the way we want to do
things going forward, but maybe its useful.