Split up queries in time
An experiment to see if it helps for certain limited "backwards" (sort by
timestamp DESC) queries, if we start with a smaller time range at the
end of the timespan and move backwards in time until we have enough
results.

This avoids grabbing all matching results and then having to sort the
whole lot in memory just to return a small chunk of them.

This reverts commit 9314f86a3b097f0645fd8ec4f73900d2ca501db8.