Add 'insert_distributed_sync=True' to Consumer (inserter) ClickHouse settings.
This setting is required so that we assured the write will be available
on the first replica (via 'load_balancing': 'in_order') after the write
returns, so post-processing can query the written data.

It needs to be set here because we *don't* want to enforce this setting
on a global level until (at minimum) this bug is fixed: https://github.com/yandex/ClickHouse/issues/2640

If it were set at a global level the replacer would explode since it
writes the minimal number of columns to handle deletions.