ref: Put function expression logic in one place (#276)
* ref: Put function expression logic in one place

Also add a topK convenience function so that we can pass eg. 'top3'
instead of 'topK(3)'

* Tidy up nested expression logic, and document it a bit more.

This technicaly removes the ability  to do:
    [func, [arg1], [arg2]] => func(arg1)(arg2)
but that was mostly a hack to allow for eg. topK(10)(foo)  which is now
no longer necessary as we can use the function alias top10 instead.