Datasets in Snuba
This attempts to provide an interface through which all of the common
functions of snuba (reading data from kafka into clickhouse, applying
mutations, and querying the data) can be delegated to various datasets.
This will allow us to have a common query API that can be used to grab
both events data, and other datasets in future.