fix: Don't use splitter for high offsets. (#307)
With the current implementation of the splitter, we have to query
for limit+offset results and do the offset trimming ourselves in
memory. When people request high page numbers, we end up with some
very large datasets in memory, so this caps the offset at which we
will use the splitter.