feat(issueless events): Hide group_id = 0 on events dataset (#352)
The current implementation of issueless events requires the group_id in clickhouse to be 0 (the column is not nullable and making it nullable has a high cost).
On the Sentry side issueless events have group == None. The fact the group_id is 0 is an implementation detail that should not be visible to Sentry, thus we can map this value to NULL on the Snuba interface. This happens both during ingestion and when querying.