build(prewhere): Change default prewhere settings try 2
Original PR: https://github.com/getsentry/snuba/pull/384
Revert PR: https://github.com/getsentry/snuba/pull/389

`sorted` was being called on the following:
```
[
  (1, (('isNull', ('issue',)), '!=', 1)),
  (5, ('project_id', 'IN', (2,))),
  (1, ('issue', 'IN', (2, 1))),
  (5, ('project_id', 'IN', (2,)))
]
```

Once `sorted` encounters a tie in the first element, it tries to
sort by the second element in the tuple, which caused problems.

We only care about sorting on the first element in the tuple cause
that will tell us which order to include the prewhere conditions.