chore: Make snuba run in container while testing (#399)
The tradeoff we're making here is speed of running the tests vs having a more accurate environment to perform our tests.

The tests used to run in 2 - 2.5 minutes. Now, they run in 4.5 to 5 minutes. But we do have a more accurate environment.

Worth it? I don't have a strong opinion. Let me know if you do.

SNS-191.