feat(api) Add support for column list in aggregate functions
We have a use-case in sentry where we want to use `argMax()` which
requires multiple columns to work. While this function is allowed in
`selected_columns` it is actually an aggregate value and should probably
be in `aggregations`. This expands the schema and query builder to allow
lists of columns in aggregations functions.

Refs getsentry/sentry#14321