fix: Fix query splitting by column
This fixes a bug with query splitting by column, where the set of
results would be incorrect if offset > 0.

We always need to remove the offset term when performing the second query
since we are adding various conditions that already restrict the results
to the exact set of events we want to return.