ref: Create DDLs
Allows you to have create statements and drop statements for a dataset.

This becomes particularly useful when we have multiple schemas for a dataset,
like in outcomes. With outcomes, we will have three schemas: the read schema,
write schema and the materialized view, which pipes content from the write
table to the read table.

So, instead of exposing all the schemas to the callsite while bootstrapping,
we will only expose a DDL class, which contains the CREATE TABLE and DROP
TABLE commands.