fix: Make group exclusion work with null group id (#431)
When group_id is null, (for issueless events) the clause group_id NOT INT (something) will always resolve to null, thus no event can be returned by any query. 
This is a hack to go around that problem as long as we do not use the transactions table for those events.