feat(cdc) Introduce the stateful consumer (#419)
For cdc we will need the consumer to show a complex stateful behavior during snapshot and in case of recovery of a snapshot.
This PR introduces an implementation of the State pattern to capture different states of the consumer. Each state is a class and it encapsulates the state specific logic.
States are mostly empty here still.