feat(cdc) Catch up state to make the snapshot consistent (#442)
When the snapshot consumer starts up, it starts in bootstrap mode. Here it consumes the cdc control topic. Once this is done we know whether we need to catch up on the WAL till the snapshot transaciton ids.

This PR implements this last step by consuming the WAL and deciding whether to apply the transactions or not depending whether the xid is included in the snapshot defined by xmin, xmax and xip_list.