feat(datasets): Break circular dependencies in preparation for query object (#461)
While building #456, it turned out I had to move a lot of code around just to break circular dependencies. That made the size of the PR explode and it made it confusing to review since it was hard to tell what was just being moved around and what actually changed.