fix(datasets): Do not run by mistake events migrations on the other datasets (#464)
Running a query (in a development environment) on a dataset other than events still runs the migraitons hardcoded for events. This breaks the table definition for anything other than events.
This moves the migration statement generation into the schema, so that they will be executed only for the dataset that defines them.