ref: Give datasets the ability to choose which rate limits they'd like (#493)
## The problem
In production, we do two types of rate limiting: global and project based. Global rate limiting is the maximum number of queries we can have at any given time. Project based rate limiting is the maximum number of queries for a particular projects. This works for any dataset that has the `ProjectExtension`. However, `outcomes` might get queries that are not scoped by project - they may be scoped by organization.

So, we want to have project based rate limits for some datasets and not for other datasets.

## Solution
The proposed solution gives us the following:
* Rate limits provided by the project extension
* Early exit if one rate limit fails