Split more modules (#514)
Adding more features to Query structure and query processing makes the Query class more complex (and this is intended). In this scenario the basic util module cannot depend on Query, because that means everyone who tries to import basic functions like to_list and escape_alias would have to depend (indirectly) on Query, which seems a wrong design.
On the other hand, moving a lot of parsing functionalities to Query means Query will have to depends on some utils.
Thus this removes the dependency from util to Query by splitting out the column parsing functions in their own module.