fix(processing) Skip invalid transaction events
If we receive events that lack `contexts.trace` we should skip those
events as they are invalid. The current event validation does not catch
transaction events that are lacking trace context so we do it here for
now.

I've not handled the trace context being malformed as that hasn't
happened _yet_ and the better solution is to fix event validation.