fix(apm) De-materialize duration column (#545)
We ran into a problem with Clickhouse materialized column. Somehow, on distributed tables, the materialized column duration does not work anymore. 0 is being explicitly passed to the local tables. We don't know why, so we are just turning it into an ordinary column.

Also it seems our consumer is causing a major issue with the fabric of space time (probably it created a wormhole) and the result is that we observed a relevant amount of transactions that finished before they started. Adding a metric to log if the issue is with related with the content of the transaction event.