feat: Introduce the AllEvents dataset
This is an experimental dataset that simply routes a query to either
the events or transactions datasets. It aligns the columns of the 2
datasets by adding a WITH clause with the missing columns.

This is an intermediate step that allows us to move the dataset
detection logic that switches between the events and transactions table
from Sentry into Snuba. In future, this will likely be solved via a
materialized view instead.