feat(datasets) Replace column method in the Query class (#555)
Now that we have query processors we need some easier way to actually tweak the query itself.
This PR introduces a method in the Query class that replaces all (well ... almost all) occurrences of a column with another one.
This is supposed to be used by query processors.

It reflects the behavior of get_all_referenced columns (which does not return all referenced columns but it is missing a few). At this point I believe it is more useful to remain consistent with that method than introducing a different behavior. This will be all fixed as soon as we have a better column abstraction.