feat: Introduce the Discover dataset (#554)
This is an experimental dataset that simply routes a query to either
the events or transactions datasets. It maps common columns to the
same name/type and fills non common columns with null.

This is an intermediate step that allows us to move the dataset
detection logic that switches between the events and transactions table
from Sentry into Snuba. In future, this will likely be solved via a
materialized view instead.