Basic classes for Query data model (#557)
The first classes needed to provide the Snuba query class with a proper AST.
This PR represents the Query as a tree. The root is the query. Nodes can be containers
(like selected columns), expressions (simple or nested) or conditions (simple or nested).

The basic idea is to provide three ways to manipulate the Query:

One is a simple iterator to go through all the nodes of a given type (like expressions). This is meant to do whatever we do today with get_all_referenced_columns
The second is mapping. We would run the mapping function over all the nodes of a given type to replace them (think about replacing columns)