feat(query) Prepare boilerplate for AST query formatting (#582)
The formatting logic for the new Snuba Query with AST will be radically different than the previous version. I will need the two systems to coexist, so this one adds some boilerplate to facilitate the process.

Now there is a second ClickhouseQuery class (ClickhouseASTQuery) that is built starting from the Snuba Query (after processing) and RequestSettings. This copies over the content of the query and adds Clickhouse specific fields. This allows us to run some Clickhosue specific query processing on the line of the query processing we do on the Snuba Query.