fix: Generate KafkaMessageMetadata in test endpoint (#593)
the test endpoint tests/DATASET/insert does not pass a KafkaMessageMetadata to the message processor, thus it breaks for CDC (which needs the offset).

This is a quick fix without changing the schema of that API call. I think we can make it more sophisticated if we wanted to let the client pass the offset, but this seems overkill for this need.