feat(query formatter) Add a Visitor to do Clickhouse specific formatting of Expressions (#594)
This is the first step into the new Query AST formatting. This PR provides what we need to format Expressions. The next step will put them together into Query formatting.

In order to make formatting Clickhouse specific, this introduces a Visitor pattern (commonly used to evaluate asts) that is used to traverse the AST and format it node by node.
This formatter also manages escaping of each node type.