fix(discover): group_id defaults to 0 instead of NULL
For the discover dataset, make group_id a non nullable column and default to 0
for transactions. This aligns group_id in Discover with the value of
transactions written to the events table.

The primary motivation for this is to prevent error like:
`DB::Exception: Type mismatch in IN or VALUES section. Expected: Nothing.
Got: UInt64.` occuring when a condition `type = "transaction" and
group_id in (1, 2, 3)` is sent to the Discover dataset.

This is an example of this error logged on Sentry.
https://sentry.io/organizations/sentry/issues/1235234594/events/2d859e6865504a36a32dc1886c381497/?project=300688

Whilst these conditions don't necessarily make sense (should always
yield an empty set since transactions have no group_id) they are
sometimes generated from Discover v2.

An alternative solution might be to strip the group_id condition
instead, however it might be more complex to implment correctly in
every case.