feat(datasets) Make complex_column_expr parametric to reuse it for the AST (#619)
complex_column_expr parses nested expressions in the current system. In the current system parsing and formatting happen together.
I think we should preserve the parsing part of it (that is dependent on the syntax) and change the outcome of that function.

This PR makes complex_column_expr parametric, it now takes three functions as parameters that are used to produce the output. It also adds a rudimentary implementation to generate Expressions to prove the concept.