perf(tagsmap): Run the flattened tags columns optimizer only for recent queries (#645)
We introduced the flattened columns optimizer on dec 11. Which means we have valid data for the flattened columns since that day.
This PR adds a condition to the optimizer so that we won't need 3 months of data to start using the optimizer. It automatically apply the optimizer only to queries that have start_date after dec 11.