feat: Customize settings when running subscription queries
This introduces `SubscriptionRequestSettings`, which overrides various settings to be hardcoded
when running subscription queries. This allows us to ignore all rate limits, and force turbo off
and consistent on. Just returning no rate limits is enough to not perform any ratelimiting, and
doesn't hit Redis to do so since we never make any calls to `rate_limit`.

I'm not totally sure if consistent is what we want, since this apparently puts a higher load on
Clickhouse, but this was mentioned previously so I've gone ahead with it for now.

This might not be the best way to handle this, but it is fairly simple in terms of code change.
Happy to work towards something more general if that's what people prefer.

Will update this with tests/docstrings if this approach looks good.