feat(query processor) Parse conditions (#653)
This takes the same approach of complex_column_expr to preserve the current parser logic. It takes conditions_expr (that essentially defines lexing and parsing for conditions), it makes that function parametric so that it can be used to build any type of expressions.

It also provides a function parse_conditions_to_expr that parses the Snuba conditions into a FunctionCall.