feat: Improve Discover column selection (#657)
If no dataset specific condition exists on a query use the other
referenced columns to help choose the dataset.

This helps ensure we return more meaningful results when the detected
dataset is ambiguous. This comes up on the organization_events_stats
endpoint in Sentry where no stats would be displayed if "duration",
or "p95" was selected as the y Axis (and no conditions were provided),
as the transactions dataset would not otherwise be selected.