perf: Support flattened tags in the discover dataset (#661)
The flattened tags column has been used in the transactions dataset for a while. This adds the support to the discover dataset.

Specifically what this PR introduces is the ability of a query processor to delegate further processing to followup processors to make the dependency explicit instead of making it hidden.
Now this support is in the discover dataset, making it generic is a TODO and requires the DatasetSelector to be made generic first.