feat: Customize settings when running subscription queries (#648)
This abstracts settings to all inherit from RequestSettings, renames the existing
RequestSettings to HTTPRequestSettings and introduces SubscriptionRequestSettings,
which overrides various settings to be hardcoded when running subscription queries.
This allows us to ignore all rate limits, and force turbo off and consistent on. Just returning
no rate limits is enough to not perform any ratelimiting, and doesn't hit Redis to do so since
we never make any calls to rate_limit.