Fix/parse function expressions (#670)
snuba.util.functions_expr contains some ad hoc tweaking for some functions that can be passed already in clickhouse format. This applies those tweaks through a query processor to the AST query.
The following step will deal with apdex and topK but it needs to do some parsing of the clickhouse expresison coming from the client.