feat(query) Parse clickhouse expressions provided in aggregations (#678)
Aggregations is the only place in the Snuba syntax where basically we allow freeform Clickhouse expressions in. This is not great, but in order to preserve backwards compatibility the query parser needs to understand how to deal with those expressions and make them available for processing.

This PR adds support for only the basics function structure to preserve backwards compatibility but this behavior is deprecated.