Update Subscription creation API stub to actually create Subscriptions
This adds a `SubscriptionCreator` class that handles the domain logic of validating and creating a
subscription. Updates the endpoint to use this and return the id.

Adds a Partitioner to determine where we'll store a given subscription. At the moment this is
hardcoded, but in the future we can update it to fetch from the dataset.