fix(metrics): Allow not having dogstatsd
All the code assumed a valid and running dogstatsd instance which was causing issues when running without it, especially with the bare Docker image over at getsentry/onpremise. This patch adds the option to have `settings.DOGSTATSD_HOST` and `settings.DOGSTATSD_PORT` as `None`, sets these to `None` by default in the Docker settings and refactors some code to use the settings values for dogstatsd at all times.

BREAKING: This removes the `--dogstatsd-host` and `--dogstatsd-port` CLI arguments, which doesn't seem to be used anywhere (checked getsentry/single-tenant and getsentry/ops).