perf(json) Test rapid json test (#719)
This allows us to replace the simplejson library with rapidjson during either deserialization in the consumer or serialization in the writer.

This is just a test to run it on specific instances of a consumer, we would never pass the option as a CLI parameter. The PR will be reverted after the test.