build(gcb): Run tests on GCB before pushing the image to Docker Hub
This required some small configuration changes on the tests side
but they are minimal. If this works out well, I'll be switching Travis
runs to use this setup too. The major difference is the lack of Redis
Cluster as it is a pain to set it up using Docker Compose. I can get it
done if it is a hard requirement for tests though.