perf(tags): Reduce the size of the arrayJoin when the client unpacks a set of tags (#707)
Snuba functions like tags_key and tags_value, unpacks the tags column and run an arrayJoin on the result to be able to GROUP BY tags. Most of the time those queries are also filtering on the tag names to extract a subset of them. The filtering has to happen after the arrayJoin has been done.

It appears that we can have a solid performance improvements if we prefilter the list of tags before the arrayJoin when the number of tags is small.