perf(prewhere) Tests adding type in the prewhere clause for selected projects (#741)
This is a test to see the results of using type in the prewhere clause. In order to do that we keep a list of suitable projects in state.
We parse the query to find project_id conditions, if at least one matches we add the custom key to the prewhere clauses.