feat(replacers) Split replacer into ReplacementWorker and ReplacementProcessor (#781)
This PR splits replacer.py into worker and processor as we did for message processors long ago. This way we can associate the replacer processor to the table writer (the storage layer). This is declared in the dataset. Which can instantiate it with all the schema specific parameters that are needed.
If/when we will build a generic replacer, we will be able to build on top of this.