feat: Record queries and timings only once for each query sent to Snuba (#739)
Record each query sent to Snuba instead of each raw query executed on
Clickhouse. Each query contains "query_list" which is a list of the
metadata associated with each query performed on Clickhouse.

Since split queries may run smaller several queries to optimize
performance, this could have resulted in the count of actual queries
recorded to be overstated and the timings associated with each to be
understated.

Recorded queries are not yet being surfaced anywhere.

This also updates metrics to be sent only once per Snuba query.