feat: Internal only datasets
Provide a way to register datasets as internal only. This prevents them
from being queried directly via the API by Sentry (and the web UI by
extension). The internal datasets can still be queried indirectly.

This is specifically for the (future) querylog dataset but it might also
apply to other datasets.