feat: Register every storage (#852)
In future, we want to be able to make Clickhouse host settings
configurable per storage (not per dataset as previously discussed).
In order to do this, we introduce a registry of storages. Future
configuration for each storage will be able to be set based on the
storage key.

This PR doesn't currently register "every" storage yet like the title says,
however that is the goal.