feat: Introduce clusters
Introduce the cluster class. The cluster provides a reader
and clickhouse connections shared by all storages that are
located on the cluster.

This PR does not yet change the settings format in any way and hardcodes
a single global cluster (essentially the same as what we had before).
This will be addedin future.