feat(cdc) Pre filter cdc messages based on table header [take 2] (#862)
The goal is to drop cdc messages before the json payload gets parsed if the Kafka headers contain a table the consumer is not interested into.

The current Consumer is not doing any dataset specific operation before json parsing and I would not want to move parsing into the MessageProcessor (since this would make the processor Kafka specific while now they are fairly agnostic to the stream type).
I added, instead, a pre-filter component (provided by the StreamLoader (which is provided by the Table Writer) together with the MessageProcessor). This is Kafka specific, if the stream is a Kafka stream, and it makes a step in making the MessageProcessor fully agnostic to the stream type.