ref: Storages are CDC aware (#872)
This moves the CDC information into the storage (rather than the
dataset). Since CDC loads data directly to a table, it seems unnecessary
to go through dataset to achieve this.

The other motivation for this is it will then allow consumers to be run
on storages rather than datasets as well.