feat(storage): Clickhouse Query placeholder (#880)
This is the first step in splitting logical query and storage (clickhouse) query.

It does the bare minimum to address two problems that I think would become much bigger if we start working together to the query planning: the query class used after storage selection is still the logical Snuba query, and the instance of the query object is still the same.
So this PR does the following:

Introduce the Clickhouse Query class (which is really just a type alias to the old class). This allows every interface in the query processing pipeline not to depend on the logical query object after the storage has been selected.

It copies the query during the translation phase, so that after storage selection we are not using the same mutable object anymore. Any change to the logical query at that point won't produce any effect.