feat(clusters): Cluster provides HTTP writer
The cluster now provides the HTTP writer based on the host/port
configuration for that cluster. This was previously based on the
global CLICKHOUSE_HOST and CLICKHOUSE_HTTP_PORT values.

The table writer is now instantiated by the writable storage since the
storage set key is shared and determines the cluster configuration that
will apply to the writer. As a result, I needed to remove most of the
subclasses of table writer and pass various writer options into the storage
instead.