feat(clusters): Map cluster to ClickHouse cluster
Clusters in Snuba can now be mapped to a physical ClickHouse cluster,
and they now aware of the set of server nodes that make up the cluster.

The cluster definition provided via settings (currently hardcoded to one
global cluster but soon to be configurable and allow more than one) will
take two additional parameters:
- distributed: Distributed vs local mode will no longer be a global
setting but a configurable property of a cluster. This allows us to use
a mix of local and distributed tables in production for different use
cases.
- cluster_name: This applies only to clusters with distributed = True,
and needs to match the ClickHouse cluster name defined in
/etc/clickhouse-server/config.xml.

The cluster exposes a get_nodes() method that returns all of the server
nodes of the cluster. In local mode, there is only one. In distributed
mode, the list of servers are found via the system.clusters table.

The get_nodes() method is not currently used anywhere (and should not be
referenced anywhere until we have updated settings.CLUSTERS in all relevant
environments). In the future, this will be used in order to run certain
scripts (such as migrations, optimize, cleanup) that must be performed
on every shard not just the distributed table.