fix(transactions) Do not apply post replacement consistency to transactions (#930)
Now we apply FINAL or exclude replaced group ids from a query in a query extension. This does not work because it is done at dataset level. This process should happen at storage level since several storages do not have replacer (let alone having a group_id column).

This PR moves that logic into a Clickhuose Query Processor which is applied only to events and errors.
This requires some additional work to find the list of project ids present in the query conditions: the extension takes the project ids from the extension data while now we have to find specific conditions in the query ... without AST still.

Since it would be extremely hard to test correctness in production (if we apply final or fail to do so by mistake we would not notice if none complained) this preserves the old behavior as well. It provides a killswitch to turn the new feature off. While this is off, both logics run and the results are compared to ensure correctness.