feat(discover) If duration column present default to transactions
Based on the metrics that have been collected, the major offender of the
impossible queries is the duration column. Since duration is often used in
aggregations this is a major source of problems.

To that end, unless a user explicitly selects the event type they want, if
duration is one of the columns in the query, use the transactions dataset.