ref(migration): Consolidate migration tasks
Migration tasks are currently split between the bootrap and migrate
commnands, which is rather confusing. It requires the user to run both
the bootstrap and migrate commands sequentially in order to ensure all
migrations actually run. Now only migrate needs to be run to perform all
the migrations.

This simplifies the implementation quite a bit, since we no longer need
the hack to ensure that materialized views are created after the
underlying raw tables. Instead we define the order in which storages
must be migrated. This implementation is intended to be temporary and
should be replaced by a proper migration system soon.