ref(transactions): Move discover column aliasing into transactions dataset
Like #934, but with hopefully less collateral damage.

This is intended to ensure that aggregation and filtering behavior is
consistent between the two different dataset representations of a
``Transaction`` (until there is a single logical representation which
will come along with the entity work) since the ``transactions`` dataset
will be the entrypoint for subscriptions on transaction data (until that
also is changed as a side effect of the entity work.)

Since this consistency is not structurally enforced, we will have to be
careful to ensure to not invalidate this property until the
implementation of entities enforces it inherently.