fix(sessions): Count abnormal sessions also as errored
Health sessions are supposed to be `sessions - errored` and errored
but not crashed/abnormal sessions are supposed to be `errored - crashed
- abnormal`.  We're currently not forcing abnormal sessions to have an
error count so the total count can be wrong.

Before we did this modification for crashed in relay but it makes more
sense to do this in snuba instead.