ref(discover) Add an error_rate function (#959)
The error_rate function is getting used more and more by the product, and it is
currently calculated on the sentry side instead of in snuba. Move the logic into
snuba so that we can properly use the AST and avoid edge case errors where the
events table is selected instead of the transactions table.