feat(ast) Better (I think) Column structure (#963)
It swaps table and column name because it seemed counterintuitive to see the table name between the column name and the path. (table name was declared after column name because, in a pre-entity world, that field would be rarely populated).

Basically everything interesting in this PR is in the query/expressions.py file. and in the formatter.py file. Everything else is just a massive codemod to swap column name and table name.