feat(AST) Promote tags in a query processor (#955)
Adds the logic to promote tags expression into promoted columns in a query processor. This is supposed to run towards the end of the processing of each storage that supports tags/contexts promotion.
The idea of running it late is that so that other processors do not have to deal with two possible coexisting representations for a tag (promoted and not promoted).

The spec (the mapping between tag keys and promoted columns) is provided in the constructor.
This wires it up to events and errors but nothing is actually happening in prod because none is yet translating the expression tags[sometthing into the function call to arrayElement.