feat(AST) Wire up the translator in the events dataset (#978)
Provides translation rules for the Events dataset.
This is the dataset with the simplest translation rules (that are not jsut default) to test the entire flow end to end. There is only one entity and one storage in such dataset so there is nothing to choose from, the rules can be passed to the query plan builder directly.