feat(ast): Wire up transactions translator (#985)
Like #978, it provides translation rules for the transactions dataset.
In order to do that it also adds a bunch of mappers that I think can be used in other scenarios as well:

column to literal (when we want to hardcode a column to a string - like the type field for transactions)
column to function
column to tag (resolves a column into a tag/context).
It also fixes a bug in managing the aliases in the previous mappers. More in a comment innline.