feat(clusters): Map cluster to ClickHouse cluster (#928)
Clusters in Snuba can now be mapped to a physical ClickHouse cluster,
and they now aware of the set of server nodes that make up the cluster.

The cluster definition provided via settings (currently hardcoded to one
global cluster but soon to be configurable and allow more than one) will
take three additional parameters:

single_node: Distributed vs local mode will no longer be a global
setting but a configurable property of a cluster. This allows us to use
a mix of local and distributed tables in production for different use
cases.
cluster_name: This applies only to clusters with single_node = False,
and needs to match the ClickHouse cluster name defined in
/etc/clickhouse-server/config.xml.
distributed_cluster_name: Also only applies to clusters with single_node
= False. If provided, the distributed hosts need to be registered as a
separate cluster in /etc/clickhouse-server/config.xml matching this name.
The cluster exposes the get_local_nodes() and get_distributed_nodes() methods
that provides a list of all of the server nodes containing the local and distributed
tables. In local mode, there is only one local node and no distributed nodes.
In distributed mode, the list of servers is found via the system.clusters table
where they must be registered.

These new methods are not currently used anywhere (and should not be
referenced anywhere until we have updated settings.CLUSTERS in all relevant
environments). In the future, this will be used in order to run certain
scripts (such as migrations, optimize, cleanup) that must be performed
on every shard not just the distributed table.