fix(schema) Enforce the valid structure of IS NULL conditions in json schema (#1010)
This enforces the conditions is valid in the request schema instead of getting to the parser.
So if a condition has a unary operator, only null is allowed as right hand side. We cannot just remove the right hand side because that would be a breaking change.
If a condition has a right hand side literal, then the operator must be binary.