fix(ast) Adapt the sql string checks in test_api to the ast query. (#1011)
The formatted ast query looks a bit different in some cases with respect to the legacy query. Specifically this impacts the conditions, the ast uses clickhouse functions.
Unfortunately there are some tests in test_api that look for expressions to be present in the sql string. That's not a great way to test to begin with, but, as long as they exist, we still need to make those tests work for the ast query.
So this PR adds ast specific conditions to those test to maintain the same behavior of the previous test and to make them work on both legacy and ast.